requirements.txt
connexion[swagger-ui]

connexion~=2.5
gunicorn~=20.0.4
werkzeug~=0.15.1
Flask~=1.1.1
Flask-Caching~=1.8.0
Flask-Migrate~=2.5.2
Flask-Script~=2.0.6
Flask-Testing~=0.8.0
pytest-factoryboy~=2.0.3
healthcheck~=1.3.3
psycopg2~=2.8.4
sqlalchemy==1.3.18
sqlalchemy_utils==0.36.7
pyjwt~=1.7.1
cryptography~=2.8
marshmallow~=3.3.0
marshmallow-sqlalchemy~=0.21.0
toolz~=0.10.0
sendgrid~=6.4.1
google-cloud-storage~=1.30.0

parameterized~=0.7.1

bcrypt~=3.1.7
six~=1.14.0
python-jose[cryptography]
